+++ tests/testthat/test_DWD_stuff.R
@@ -4,20 +4,98 @@ test_that("download of list of stations works", {
   # don't run download tests on cran, travis, appveyor
   skip_on_cran(); skip_on_travis(); skip_on_appveyor()
 
+  # Structure
+  # .
+  # ├── daily
+  # │   ├── climate
+  # │   │   ├── historic
+  # │   │   └── recent
+  # │   └── precip
+  # │       ├── historic
+  # │       └── recent
+  # └── monthly
+  #     ├── climate
+  #     │   ├── historic
+  #     │   └── recent
+  #     └── precip
+  #         ├── historic
+  #         └── recent
 
-  # stations with daily climate data from the last year
+  # Test some of the possible lists not all.
+
+  # Column widths, and thus parsing, is different for daily and monthly lists.
+  # All monthly station lists have floating point elev.
+  # A data.frame with 8 columns ist constructed during reading/parsing.
+  # If parsing went well, the columns should have proper types.
+
+  # DAILY =====================================================================
+  # daily climate recent ------------------------------------------------------
   expect_silent(stations <- read.DWDstations())
   expect_is(stations, 'data.frame')
   expect_length(stations, 8)
   expect_gt(nrow(stations), 1000)
-  expect_named(stations,
-               c("id", "from", "to", "elev", "lat", "long", "name", "state"))
 
-  # precipitation stations with monthly historical data
-  expect_silent(precip <- read.DWDstations(type='precip', period='historical',
-                                            resolution='monthly'))
+  expect_is(stations$id,   'integer')
+  expect_is(stations$from, 'Date')
+  expect_is(stations$to,   'Date')
+  expect_is(stations$elev, 'integer')
+  expect_is(stations$lat,  'numeric')
+  expect_is(stations$long, 'numeric')
+
+  # daily climate historic ----------------------------------------------------
+  expect_silent(stations <- read.DWDstations(period='historic'))
+  expect_is(stations, 'data.frame')
+  expect_length(stations, 8)
+  expect_gt(nrow(stations), 1000)
+
+  expect_is(stations$id,   'integer')
+  expect_is(stations$from, 'Date')
+  expect_is(stations$to,   'Date')
+  expect_is(stations$elev, 'integer')
+  expect_is(stations$lat,  'numeric')
+  expect_is(stations$long, 'numeric')
+
+  # daily precip recent -------------------------------------------------------
+  expect_silent(precip <- read.DWDstations(type='precip'))
   expect_is(precip, 'data.frame')
+  expect_length(precip, 8)
   expect_gt(nrow(precip), 3000)
+
+  expect_is(precip$id,   'integer')
+  expect_is(precip$from, 'Date')
+  expect_is(precip$to,   'Date')
+  expect_is(precip$elev, 'integer')
+  expect_is(precip$lat,  'numeric')
+  expect_is(precip$long, 'numeric')
+
+
+  # MONTHLY ===================================================================
+  # monthly climate recent ----------------------------------------------------
+  expect_silent(stations <- read.DWDstations(resolution='monthly'))
+  expect_is(stations, 'data.frame')
+  expect_length(stations, 8)
+  expect_gt(nrow(stations), 1000)
+
+  expect_is(stations$id,   'integer')
+  expect_is(stations$from, 'Date')
+  expect_is(stations$to,   'Date')
+  expect_is(stations$elev, 'numeric')
+  expect_is(stations$lat,  'numeric')
+  expect_is(stations$long, 'numeric')
+
+  # monthly precip historic ---------------------------------------------------
+  expect_silent(precip <- read.DWDstations(type='precip', period='historic',
+                                           resolution='monthly'))
+  expect_is(precip, 'data.frame')
+  expect_length(precip, 8)
+  expect_gt(nrow(precip), 3000)
+
+  expect_is(precip$id,   'integer')
+  expect_is(precip$from, 'Date')
+  expect_is(precip$to,   'Date')
+  expect_is(precip$elev, 'numeric')
+  expect_is(precip$lat,  'numeric')
+  expect_is(precip$long, 'numeric')
 })